Factor a general remote FS accessor out of BinaryCacheStore
